requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.2.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
fonttools==4.53.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pymongo==4.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.3.0
Werkzeug==3.1.3
